serverside/requirements.txt
Django==1.6.5
Markdown==2.4.1
MySQL-python==1.2.3
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
apt-xapian-index==0.45
argparse==1.2.1
billiard==3.3.0.17
celery==3.1.11
chardet==2.0.1
colorama==0.2.5
django-celery==3.1.10
django-cors-headers==0.12
django-filter==0.7
django-flashpolicies==1.5
djangorestframework==2.3.13
html5lib==0.999
kombu==3.0.18
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
pytz==2014.4
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
urllib3==1.7.1
wsgiref==0.1.2
